requirements.txt
opencv-python>=4.8.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0
pytest>=7.3.1
pytest-asyncio>=0.23.0
